docs: update checked in tutorial lib based on tutorial steps
