+++ .tutorial/tests/integration/charms/actions.yaml
@@ -2,3 +2,4 @@
 # consider adding an action for each thing you want to test
 
 lib-version:
+charm-hostname:
